Move all utils to their own files
