+++ pajbot/utils/dump_threads.py
@@ -1,0 +1,11 @@
+import sys
+
+
+def dump_threads():
+    import threading
+    import traceback
+
+    for th in threading.enumerate():
+        print(th)
+        traceback.print_stack(sys._current_frames()[th.ident])
+        print()
