+++ pajbot/utils/print_traceback.py
@@ -1,0 +1,5 @@
+import traceback
+
+
+def print_traceback():
+    traceback.print_stack()
