+++ pajbot/utils/remove_none_values.py
@@ -1,0 +1,2 @@
+def remove_none_values(d):
+    return {k: v for k, v in d.items() if v is not None}
