+++ pajbot/utils/time_limit.py
@@ -1,0 +1,17 @@
+import signal
+from contextlib import contextmanager
+
+from pajbot.exc import TimeoutException
+
+
+@contextmanager
+def time_limit(seconds):
+    def signal_handler(signum, frame):
+        raise TimeoutException("Timed out!")
+
+    signal.signal(signal.SIGALRM, signal_handler)
+    signal.alarm(seconds)
+    try:
+        yield
+    finally:
+        signal.alarm(0)
